+++ .env.example
@@ -1,0 +1,2 @@
+PG_PASS=password
+AUTHENTIK_SECRET_KEY=password
